+++ includes/class-testimonials-widget-archives-widget.php
@@ -1,7 +1,7 @@
 <?php
 /**
 Testimonials Widget
-Copyright (C) 2015 Michael Cannon
+Copyright (C) 2015 Axelerant
 
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by